# docs/swagger.yaml
basePath: /api
definitions:
  common.Page:
    properties:
      current:
        type: integer
      list: {}
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  common.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  define.CreatePermissionRequest:
    properties:
      action:
        type: string
      description:
        type: string
      name:
        type: string
      target:
        type: string
    required:
    - action
    - name
    - target
    type: object
...
  define.CreateTenantRequest:
    properties:
      domain:
        type: string
      name:
        type: string
    required:
    - domain
    - name
    type: object
...
  define.ListTenantPageResponse:
    properties:
      createdAt:
        type: string
      domain:
        type: string
      name:
        type: string
      tenantId:
        type: integer
      updatedAt:
        type: string
    type: object
...
  define.LoginByUsernameRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  define.MyInfoResponse:
    properties:
      avatar:
        type: string
      email:
        type: string
      mobile:
        type: string
      userId:
        type: integer
      username:
        type: string
    type: object
  define.PermissionPageResponse:
    properties:
      action:
        type: string
      description:
        type: string
      name:
        type: string
      permissionId:
        type: integer
      target:
        type: string
    type: object
  define.RegisterUserRequest:
    properties:
      email:
        type: string
      password:
        maxLength: 20
        minLength: 8
        type: string
      username:
        maxLength: 20
        minLength: 6
        type: string
    required:
    - email
    - password
    - username
    type: object
...
  define.TokenResponse:
    properties:
      accessToken:
        type: string
      expireTime:
        type: integer
    type: object
  define.UpdateMyInfoRequest:
    properties:
      avatar:
        type: string
    type: object
  define.UpdateMyPasswordRequest:
    properties:
      newPassword:
        type: string
      oldPassword:
        type: string
      repeatPassword:
        type: string
    required:
    - newPassword
    - oldPassword
    - repeatPassword
    type: object
  define.UpdatePermissionRequest:
    properties:
      action:
        type: string
      description:
        type: string
      name:
        type: string
      permissionId:
        type: integer
      target:
        type: string
    required:
    - permissionId
    type: object
...
  define.UpdateTenantRequest:
    properties:
      domain:
        type: string
      name:
        type: string
      tenantId:
        type: integer
    required:
    - tenantId
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a ether api server.
  title: Ether API
  version: "1"
paths:
  /auth/loginByUsername:
    post:
      consumes:
      - application/json
      description: Login by username
      parameters:
      - description: LoginByUsernameRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/define.LoginByUsernameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/define.TokenResponse'
              type: object
      summary: Login by username
      tags:
      - auth
...
  /auth/registerByEmail:
    post:
      consumes:
      - application/json
      description: Register by email
      parameters:
      - description: RegisterUserRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/define.RegisterUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      summary: Register by email
      tags:
      - auth
  /my:
    get:
      consumes:
      - application/json
      description: Get my info
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/define.MyInfoResponse'
              type: object
      security:
      - Bearer: []
      summary: Get my info
      tags:
      - my
    put:
      consumes:
      - application/json
      description: Update my info
      parameters:
      - description: UpdateMyInfoRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/define.UpdateMyInfoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - Bearer: []
      summary: Update my info
      tags:
      - my
  /my/password:
    put:
      consumes:
      - application/json
      description: Update my password
      parameters:
      - description: UpdateMyPasswordRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/define.UpdateMyPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - Bearer: []
      summary: Update my password
      tags:
      - my
  /permission:
    get:
      consumes:
      - application/json
      description: List permissions
      parameters:
      - description: current
        in: query
        name: current
        type: integer
      - description: page_size
        in: query
        name: page_size
        type: integer
      - description: name
        in: query
        name: target
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/common.Page'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/define.PermissionPageResponse'
                        type: array
                    type: object
              type: object
      security:
      - Bearer: []
      summary: List permissions
      tags:
      - permission
    post:
      consumes:
      - application/json
      description: Create permission
      parameters:
      - description: CreatePermissionRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/define.CreatePermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - Bearer: []
      summary: Create permission
      tags:
      - permission
  /permission/{permission_id}:
    delete:
      consumes:
      - application/json
      description: Delete permission
      parameters:
      - description: permission_id
        in: path
        name: permission_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - Bearer: []
      summary: Delete permission
      tags:
      - permission
    put:
      consumes:
      - application/json
      description: Update permission
      parameters:
      - description: permission_id
        in: path
        name: permission_id
        required: true
        type: integer
      - description: UpdatePermissionRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/define.UpdatePermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - Bearer: []
      summary: Update permission
      tags:
      - permission
...
  /tenants:
    get:
      consumes:
      - application/json
      description: List tenants
      parameters:
      - in: query
        name: current
        type: integer
      - in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/common.Page'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/define.ListTenantPageResponse'
                        type: array
                    type: object
              type: object
      security:
      - Bearer: []
      summary: List tenants
      tags:
      - tenant
    post:
      consumes:
      - application/json
      description: Create tenant
      parameters:
      - description: CreateTenantRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/define.CreateTenantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - Bearer: []
      summary: Create tenant
      tags:
      - tenant
  /tenants/{tenantId}:
    delete:
      consumes:
      - application/json
      description: Delete tenant
      parameters:
      - description: tenant_id
        in: path
        name: tenant_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - Bearer: []
      summary: Delete tenant
      tags:
      - tenant
    put:
      consumes:
      - application/json
      description: Update tenant
      parameters:
      - description: tenant_id
        in: path
        name: tenant_id
        required: true
        type: integer
      - description: UpdateTenantRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/define.UpdateTenantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - Bearer: []
      summary: Update tenant
      tags:
      - tenant
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"